% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{cov.huber}
\alias{cov.huber}
\title{Covariance estimation with Huber's loss function}
\usage{
cov.huber(X)
}
\arguments{
\item{X}{a n x p data matrix with each row being a sample.}
}
\value{
A list with the following items
\item{covhat}{the vector of estimated means}
}
\description{
This function estimates covariance of multivariate data using the Huber's loss. The tuning parameter is chosen by cross validation.
}
\examples{
set.seed(100)
p = 20
n = 10
X = matrix(rnorm( p*n, 0,1), nrow = n)
covhat = cov.huber(X)

}
\references{
Huber, P.J. (1964). "Robust Estimation of a Location Parameter." The Annals of Mathematical Statistics, 35, 73–101.
}
