% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{farm.loadings}
\alias{farm.loadings}
\title{Estimate Factor loadings and factors}
\usage{
farm.loadings(object)
}
\arguments{
\item{object}{an object produced from the farm.scree function}
}
\value{
A list with the following items
\item{loadings}{the factor loadings}
\item{factors}{the factors}
}
\description{
This function estimates the factor (loadings) given the estimated number of factors and covariance matrix. These can be found as output to the \code{\link{farm.scree}} function.
}
\examples{
set.seed(100)
p = 20
n = 10
X = matrix(rnorm( p*n, 0,1), nrow = n)
output  = farm.scree(X)
loadings = farm.loadings(output)
}
