% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{farm.scree}
\alias{farm.scree}
\title{Diagnostic plots needed to determine the number of underlying factors}
\usage{
farm.scree(X, K.scree = NULL, K.factors = NULL, robust = FALSE)
}
\arguments{
\item{X}{an n x p data matrix with each row being a sample.}

\item{K.scree}{an \emph{optional} integer specifying the number of eigenvalues to be plotted in the scree plot. Default is min(n,p).}

\item{K.factors}{an \emph{optional} integer specifying the number of eigenvalues to be used for the eigenvalue ratio test. Default is (min(n,p)/2)}

\item{robust}{a logical indicating whether to use a robust covariance estimator if TRUE, or the sample covariance estimator. Default is FALSE.}
}
\value{
\item{Two plots:}{First plot is the scree plot of the data. Second plot illustrates the eigenvalue ratio test. }
}
\description{
Given the data, this function  draws a scree plot and a plot of the eigenvalue ratios .
}
\details{
The maximum eigenvalue ratio is marked differently on the plot.  The index of this maximum ratio gives the number of estimated factors.
details User has to hit <Return> to see the second plot.
}
\examples{
p = 50
n = 20
epsilon = matrix(rnorm( p*n, 0,1), nrow = n)
B = matrix(rnorm(p*3,0,1), nrow=p)
fx = matrix(rnorm(3*n, 0,1), nrow = n)
X = fx\%*\%t(B)+ epsilon
farm.scree(X)
}
