% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{farmTest}
\alias{farmTest}
\title{FarmTest with unknown factors}
\usage{
farmTest(X, h0, K = -1L, alpha = 0.05, alternative = "two.sided")
}
\description{
Internal function implemented in C++ for FarmTest with unknown factors. This is a case of \code{farm.test}.
}
\seealso{
\code{\link{farm.test}}
}
