% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{farmTestTwo}
\alias{farmTestTwo}
\title{Two sample FarmTest with unknown factors}
\usage{
farmTestTwo(X, Y, h0, KX = -1L, KY = -1L, alpha = 0.05,
  alternative = "two.sided")
}
\description{
Internal function implemented in C++ for two sample FarmTest with unknown factors. This is a case of \code{farm.test}.
}
\seealso{
\code{\link{farm.test}}
}
