% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{farmTestTwoFac}
\alias{farmTestTwoFac}
\title{Two sample FarmTest with known factors}
\usage{
farmTestTwoFac(X, facX, Y, facY, h0, alpha = 0.05,
  alternative = "two.sided")
}
\description{
Internal function implemented in C++ for two sample FarmTest with known factors. This is a case of \code{farm.test}.
}
\seealso{
\code{\link{farm.test}}
}
