% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{huberCov}
\alias{huberCov}
\title{Huber-type covariance estimation}
\usage{
huberCov(X, n, p)
}
\arguments{
\item{X}{An \eqn{n} by \eqn{p} data matrix.}

\item{n}{Number of rows of \code{X}.}

\item{p}{Number of columns of \code{X}.}
}
\description{
Internal function implemented in C++ for tuning-free Huber-type covariance estimation. This function is incorporated into \code{farm.cov}.
}
\seealso{
\code{\link{farm.cov}}
}
