% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{huberMean}
\alias{huberMean}
\title{Huber mean estimation}
\usage{
huberMean(X, n, epsilon = 1e-04, iteMax = 500L)
}
\description{
Internal function implemented in C++ for tuning-free Huber mean estimation. See \code{farm.mean} for details.
}
\seealso{
\code{\link{farm.mean}}
}
