% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmTest}
\alias{rmTest}
\title{Robust multiple testing}
\usage{
rmTest(X, h0, alpha = 0.05, alternative = "two.sided")
}
\description{
Internal function implemented in C++ for robust multiple testing without factor-adjustment. This is a case of \code{farm.test}.
}
\seealso{
\code{\link{farm.test}}
}
