% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmTestTwo}
\alias{rmTestTwo}
\title{Two sample robust multiple testing}
\usage{
rmTestTwo(X, Y, h0, alpha = 0.05, alternative = "two.sided")
}
\description{
Internal function implemented in C++ for two sample robust multiple testing without factor-adjustment. This case is incorporated into \code{farm.test}.
}
\seealso{
\code{\link{farm.test}}
}
