% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{farm.FDR}
\alias{farm.FDR}
\title{Control FDR given a list of pvalues}
\usage{
farm.FDR(pvalue, alpha, type = c("mBH", "BH"), lambda = seq(0.05, 0.95,
  0.05), pi0.method = c("smoother", "bootstrap"), smooth.df = 3,
  smooth.log.pi0 = FALSE)
}
\arguments{
\item{pvalue}{a vector of p-values obtained from multiple testing}

\item{alpha}{an \emph{optional} significance level for testing (in decimals). Default is 0.05.}

\item{type}{an \emph{optional} character string specifying the type of test. The default is the modified BH procedure (type = "mBH"). The usual BH procedure is also available (type = "BH").}

\item{lambda}{an \emph{optional} threshold for estimating the proportion of true null hypotheses pi_0. Must be in [0,1).}

\item{pi0.method}{Either "smoother" or "bootstrap"; the method for automatically choosing tuning parameter in the estimation of pi_0, the proportion of true null hypotheses. Optional.}

\item{smooth.df}{Number of degrees-of-freedom to use when estimating pi_0 with a smoother. Optional.}

\item{smooth.log.pi0}{If TRUE and pi0.method = "smoother", pi_0 will be estimated by applying a smoother to a scatterplot of log(pi_0) estimates against the tuning parameter lambda. Optional.}
}
\value{
\item{rejected}{the indices of rejected hypotheses, along with their corresponding p values, and adjusted p values, ordered from most significant to least significant}
\item{alldata}{all the indices of the tested hypotheses, along with their corresponding p values, adjusted p values, and a column with 1 if declared siginificant and 0 if not}
}
\description{
Given a list of p-values, this function conducts multiple testing and outputs the indices of the rejected hypothesis. Uses an adaptive Benjamini-Hochberg (BH) procedure where the proportion of true nulls is estimated. Uses the \code{\link[qvalue]{pi0est}} function to estimate this proportion. See Storey(2015).
}
\examples{
X = rnorm( 1000, 0,1)
X = matrix( X, 100)
pval = apply(X,1, function(x) t.test(x)$p.value)
farm.FDR(pval, 0.05)
farm.FDR(pval, 0.01, type = "BH")

}
\references{
Storey JD (2015). qvalue: Q-value estimation for false discovery rate control. R package version 2.8.0, \url{http://github.com/jdstorey/qvalue.}
}
