% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmtest-package.r
\docType{package}
\name{FarmTest}
\alias{FarmTest}
\alias{FarmTest-package}
\title{FarmTest: Factor Adjusted Robust Multiple Testing}
\description{
This R package conducts multiple hypothesis testing of mean effects. It implements a robust procedure to estimate distribution parameters and accounts for strong dependence among coordinates via an approximate factor model. This method is particularly suitable for high-dimensional data when there are thousands of variables but only a small number of observations available. Moreover, the method is tailored to cases when the underlying distribution deviates from Gaussianity, which is commonly assumed in the literature.
For detailed information on how to use and install see \url{https://kbose28.github.io/FarmTest/}. See the paper on the 'FarmTest' method, Zhou et al.(2017) at https://goo.gl/68SJpd, for detailed description of methods and further references.
}
