% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{farmTestFac}
\alias{farmTestFac}
\title{FarmTest with known factors}
\usage{
farmTestFac(X, fac, h0, alpha = 0.05, alternative = "two.sided")
}
\description{
Internal function implemented in C++ for FarmTest with known factors. This is a case of \code{farm.test}.
}
\seealso{
\code{\link{farm.test}}
}
