% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{farmTestFacBoot}
\alias{farmTestFacBoot}
\title{FarmTest with known factors and multiplier bootstrap}
\usage{
farmTestFacBoot(X, fac, h0, alpha = 0.05, alternative = "two.sided",
  B = 500L)
}
\description{
Internal function implemented in C++ for FarmTest with known factors, where p-values are obtained via multiplier bootstrap. 
This is a case of \code{farm.test}.
}
\seealso{
\code{\link{farm.test}}
}
