% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRej}
\alias{getRej}
\title{Multiple testing via an adaptive Benjamini-Hochberg procedure}
\usage{
getRej(Prob, alpha, p)
}
\description{
Internal function implemented in C++ for an adaptive Benjamini-Hochberg procedure. This function is incorporated into \code{farm.fdr}.
}
\seealso{
\code{\link{farm.fdr}}
}
