% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{huberCov}
\alias{huberCov}
\title{Huber-type covariance estimation}
\usage{
huberCov(X, n, p)
}
\description{
Internal function implemented in C++ for tuning-free Huber-type covariance estimation. See \code{farm.cov} for details.
}
\seealso{
\code{\link{farm.cov}}
}
