% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{huberMean}
\alias{huberMean}
\title{Huber mean estimation}
\usage{
huberMean(X, n, epsilon = 1e-04, iteMax = 500L)
}
\arguments{
\item{X}{An \eqn{n}-dimensional data vector.}

\item{n}{The length of \code{X}.}

\item{epsilon}{An \strong{optional} numerical value for tolerance level. The default value is 0.0001.}

\item{iteMax}{An \strong{optional} integer for maximun number of iteration. The default value is 500.}
}
\description{
Internal function implemented in C++ for tuning-free Huber mean estimation. This function is incorporated into \code{farm.mean}.
}
\seealso{
\code{\link{farm.mean}}
}
