% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{print.farm.test}
\alias{print.farm.test}
\title{Print method for the class farm.test.
Not intended for end-user use.}
\usage{
\method{print}{farm.test}(obj)
}
\description{
Print method for the class farm.test.
Not intended for end-user use.
}
