% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmTestTwoBoot}
\alias{rmTestTwoBoot}
\title{Two sample robust multiple testing with multiplier bootstrap}
\usage{
rmTestTwoBoot(X, Y, h0, alpha = 0.05, alternative = "two.sided",
  B = 500L)
}
\description{
Internal function implemented in C++ for two sample robust multiple testing without factor-adjustment, where p-values are obtained via multiplier bootstrap. 
This is a case of \code{farm.test}.
}
\seealso{
\code{\link{farm.test}}
}
